requirements.txt
fsspec==2025.5.1
ftfy==6.3.1
gradio==5.35.0
huggingface_hub==0.33.2
iopath==0.1.10
lightning==2.5.2
matplotlib==3.10.3
omegaconf==2.3.0
openpyxl==3.1.5
Pillow==11.3.0
PyYAML==6.0.2
regex==2024.11.6
safetensors==0.5.3
scikit_learn==1.7.0
setuptools==60.2.0
Shapely==2.1.1
tabulate==0.9.0
timm==1.0.22
torchmetrics==1.7.3
transformers==4.47.1
hydra-core==1.3.2
numpy==1.23.5
accelerate==1.8.1
tqdm
faiss-gpu
seaborn
einops